services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.12
redis==5.0.8
prometheus-client==0.21.0
requests==2.32.3
structlog==24.4.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
google-cloud-secret-manager==2.21.0
google-cloud-kms==2.24.0
google-cloud-pubsub==2.26.0
jsonschema==4.23.0
cachetools==5.5.0
